BW-2225
Patch the _ssl standard extension module so that when we invoke
getpeercert(), the dictionary returned includes a 'pubkey' entry,
which is a DER encoding of the certificates public key.  Except it
is not quite the same DER encoding that we get out of openssl --
it is missing a 24 byte header.