Install the rootfs to our install target
The rootfs gets put in a rootfs/ directory inside our install target
so that the bundle script can make a tarball out of it.  This prevents
us from installing straight to an actual filesystem, but we can't
really do that anyway because this isn't quite a real rootfs.  It
needs the file to be owned by root and needs to have a few device
files created.

We also install a staging directory, which contains the extra files
that are needed to build against the libraries build-root provides,
but are not actually needed on the final file system.