Install files for tarball building + symlink fixes
We need both the makedevs executable that build-root creates and also
the device table that it builds in order to use fakeroot to build the
tarball.  We currently just use system fakeroot so we don't install
the fakeroot that build-root builds.

I also noticed that symlinks to directories were not getting installed
because of a misunderstanding about how os.walk deals with symlinks.
Once I fixed that I found and fixed another non relative symlink.