Lots of updates to get things building post-merge
So apparently being able to put a single defconfig file in the root
directory and select it with make defconfig was never intended behavior,
which is probably why it only worked reliably if we did it twice.
Instead we have to put a named defconfig in the configs folder and then
use that name as the make target.

Our defconfig also had a bunch of deprecated items.  udev is no longer
supported, so for now I switched to eudev.  I think eventually we should
be using systemd though.  We also have to specify a kernel version, which
has to match the kernel version of the headers packaged with our compiler.
Which raises the question of whether it is okay to compile against a 3.17
kernel and then run everything on a 3.14 kernel.

We also have to update our accumulated python hacks to work with python3.4,
and there are a couple of new places where absolute symlinks are created.
Also I accidentally deleted a few hacks that were in place to let the TR
bundle script create the final root filesystem tarball.