Update zxing to master of zxing-cpp
The cpp port of zxing was removed from the official zxing repo and is
now being maintained separately.  There are no official releases of
this branch, but it does have a cmake build setup, and comes with a
simple cli client, whereas the current zxing builds only a static
library, so nothing actually gets installed to the bot.