Backport non-in-source cmake builds for zxing
The zxing-cpp fork we are using switched from scons to cmake, but it
specifically does not work with in source builds.  Our version of build
root doesn't support this, so for now I backported this feature from a
newer release.  This still leaves in source builds as the default behavior
so we won't break other packages that use cmake and aren't written to
support non in source builds.