Add in extra wpa features for AP mode
I don't know if WPS support is also necessary, but I added it just in case.
And the wpa cli client is useful for debugging and doesn't take up too much
space.  I also passed this through savedefconfig (which we _should_ be doing
every time we commit), so things got rearranged a bit and unnecessary
entries were removed.