Tell avahi to always win hostname conflicts
This is a quick fix for the bug reported here
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=780981 in avahi and
not since fixed. It may not be exactly this bug (there's more reports of
phantom hostname conflicts bouncing around the internet) but it has the
same effects. There's two places in the core where avahi will withdraw
rrsets, if fixing/breaking this one doesn't fix it we'll do the other one.