Build kernel, ti gfx drivers and qt egl
The setup for building the ti graphics drivers is dependent on the
build-root setup for building a linux kernel because the ti graphics
driver sdk scripts need some information about the kernel they are
going to be running under.  The right way to handle this would be to
add a minimal number of files to the morepork-linux artifact in order
to provide this information and then add a way to point these scripts
at those files.  But so far the only thing I have gotten working is to
have build-root build the linux kernel locally and use that instead of
what we are building in morepork-linux.

This might be an acceptable long term solution except that it is an
incredibly shitty kernel development flow, and it doesn't actually
pick up on kernel changes -- build-root grabs the git contents once
and stores them as a tarball and then uses that tarball for every
subsequent build.