Update wifi packages to a newer build-root release
These are the connman and wpa_supplicant versions from the 2017.02.x release.
This gives us both fixes from the upstream projects and fixes to the buildroot
patches to these projects.  We should still go through the local fixes we made
for birdwing and see if we want any of those here.

There is one change from upstream -- the connman config was checking for a
compiler feature that this version does not know about.