Skip building the eglfs emulator plugin
Just removing this plugin does not work because the cmake config generated by
qt still expects this plugin to exist when we are building the ui.  Qt seems
to be configured to build this plugin whenever opengl is enabled, so patching
qt seems to be the only way to avoid this.