Patch e2fsprogs to build on newer glibc versions
Rename a static local function to no longer conflict with a glibc function
copy_file_range that only exists in relatively new glibc versions.  This has
been fixed upstream since version 1.43.8, so we _could_ also fix this by
updating e2fsprogs, but this would require a lot more testing due to the large
number of other changes in that upgrade.