Update device table name/location. Add user table.
We want to use the device table that is automatically updated by buildroot.
It's name has also actually changed since our previous version of buildroot.
Also we are adding in the user table file and the executeable to generate the
user table. This is necessary for the avahi and systemd users to be added so
these services actually start on the bot.