makerbot - add build python scripts and configs
Some noteable changes since last version update:

In mbdefconfig:
Remove NULL rootfs. By having a rootfs NULL in place, we actually do
not run any of the fakeroot commands for setting up the user and
device table files. We now want these files in place so we use them
in jenkins.py and do_package.py.
BR2_TARGET_GENERIC_GETTY_PORT was set on 'console' (the default
setting) and we need to change this to ttyS0 for newer versions on
buildroot. This was causing issues with connecting with a TTL cable
and the console input freezing. Another tell-tale sign was the login
message was printing twice.
Add zlib python3 module.

In build.py:
Update device table name/location. We want to use the device table
that is automatically updated by buildroot. It's name has also
actually changed since our previous version of buildroot.
Copy the passwd, shadow, and group files that are generated by
mkusers directly to the Install/rootfs/etc directory. This is
necessary for the avahi and systemd users to be added so these
services actually start on the bot.
We are now using mkfs.ext4 instead of mke2img in do_package. fsck
threw a lot of errors when images were built using mke2img.