Do not delete the output/build/buildroot-fs/tar dir or it's contents.
The etc shadow, group, and passwd files that are generated under the fakeroot
session are placed in this tar directory. We (potentially) need them and
therefore have to prevent this directory from being deleted. With that, we
also need all the files within the tar directory and these get deleted in the
fakeroot session in fs/common.mk.