Use static compilation for host extfs utils
We use the version of mkfs.ext4 that we compile here for generating our
final filesystems.  We have to support doing this just from the artifact
that we produce, so we put mkfs.ext4 into that artifact.  But when using
shared libraries there are 5 libraries that we would also have to copy
into the artifact and ensure that we use instead of system versions of
these libraries which are likely to be too old to use and also ensure
that we don't use the libraries in place in the staging directory here,
since on jenkins this directory could easily exist with a different,
incompatible version of these libraries.  A static compilation is a much
simpler way to bundle the full functionality of the utility into an
artifact.