Use python3 for python scripts here
Sam did some work on these scripts and included some code that only
works on python3, but left the python version as `python` which is
still python2 on most systems.  The code is unused when doing full
clean builds as jenkins does, so this wasn't caught by jenkins.  We
are trying to migrate to python3, this code has most recently been
thoroughly tested by Sam on python3, and there doesn't appear to be
any python2 only code in here.