Generate debug symbols locally
This adds a new buildroot option to generate debug symbols for all binaries
that it compiles while making no other changes to what it builds (the
existing option that claims to do this also turns on additional debugging
modes in qt5 and boost).  When combined with the buildroot feature to strip
all binaries when moving them from staging to target, this gives us debug
symbols in the staging directory while making no changes at all to what
actually gets installed to the firmware.

The change here is just in the config, creating a new option that lets a
debug level be set (which adds -g to CFLAGS and CXXFLAGS for all packages)
without setting the debug flag that triggers debug modes in qt5/boost.