Remove useless / mostly useless packages
MTD utils was only carried over from birdwing which used that as a filesystem.
Here it is completely useless.  Bzip2 compression is just not used by anything
in the firmware.  Valgrind certainly has a lot of interesting use cases, but
after trying to use it for a while it seems like we don't actually have enough
resources to use it directly in the firmware in any useful way.

The vim runtime files are a bit more of a tough call, since they do provide
syntax highlighting and some other convenietn features for serious text
editting.  But 10MB is really just too much disk space given how rare
serious text editting on the printer is (and IMO should be).

This also replaces gdb with gdb server, since one is 4MB and the other is
300kB.  I am certainly much more used to just running gdb directly, but it
isn't really any more limiting to use gdb server plus the gdb that comes
with our linaro toolchain, and also gdb is a standalone binary so it is
really easy to just scp it onto a bot to use it.