Speed up the keyboard response to capitalization
BW-5180
https://makerbot.atlassian.net/browse/BW-5180

This removes the small text from the default keyboard as it is unused
(and really can't be used unless we have a tablet sized screen).  More
importantly it changes the implemntation of key capitalization, so that
instead of changing a font property, we just have two different static
text items that we change the visibility on.

This eliminates keyboard lag in most situations.  The lag in bringing
the keyboard up for the first time is just initiailzing capitalization,
and the lag after the first letter typed comes from switching to lower
case letters.

There is still significant lag in switching between keyboard screens,
which would require additional investigation to address.