Speed up switch to/between symbol screens
Pretty much all of the lag in both switching between the alphabet and
symbol layouts and between the two symbol layouts themselves is spent in
loading and rendering the qml for these screens, because we only ever
have one of those three screens loaded at a time.  Specifically there is
one main loader object that chooses between a large number of potential
layouts (a combination of dozens of locales and a number of layout types
although I don't see any way to get at layout types other than "main"
and "symbols") and then the symbols layout (which is generally the same
for most locales) is itself a loader which loads one of two component
pages.

To fix this lag we want to have all three keyboard layouts loaded for
whatever locale we have chosen, and when we switch between these layouts
we just change which layout is visible.

For switching between the two symbol screens this is a simple change --
we just build a single KeyboardLayout with 8 rows instead of 4, changing
the visibility of the individual rows.  For switching between the main
and symbol screens we make a new "Preloader" class which attempts to be
a drop in replacement for the Loader class but which keeps one specific
source preloaded so that changing to the preloaded class is a change in
visibility settings instead of actually loading a new class.  (We only
need to keep one of the two sources preloaded because switching from the
preloaded source back to whatever source was previously selected is also
just a visibility change, and we choose the "symbols" layout so that
both this and the "main" layout will be loaded at boot up.)

The lag when switching between the "main" and "symbols" layouts drops to
just under 1 second from about 5 seconds with this change, which is good
but also still noticably longer than switching between symbol pages or
switching between capital and lower case.  Also the lag in switching
languages has increased, but only from ~6 seconds to ~11 seconds.